feat: replace manager frontend PWA with mapper frontend PWA
