+++ src/frontend/package.json
@@ -41,7 +41,6 @@
     "tailwindcss": "^3.4.3",
     "typescript": "^5.4.5",
     "vite": "^5.2.12",
-    "vite-plugin-pwa": "^0.20.0",
     "vitest": "^1.6.0",
     "workbox-window": "^7.1.0"
   },
